backend/requirements.txt
asgiref==3.11.1
Django==6.0.2
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework-camel-case==1.4.2
pillow==12.2.0
sqlparse==0.5.5
tzdata==2025.3
